Edit and buff Helioflare Power Forge
Its on kill effect was weak for what would be a massive directed solar flare laser beam, so now it absolutely destroys everything that was on that slot. I think the killed creature's equipment would be destroyed from that, plus anything that spawns cards on death. It also now has the previously missing effect of replacing the power card with pollution if it kills a player, with the same destructiveness of its main on kill effect.